+++ src/op/converse.rs
@@ -21,10 +21,10 @@
 
 /// Return a digraph's converse.
 ///
-/// # How do I implement `Converse`
+/// # Implementing [`Converse`] for a digraph type
 ///
-/// Provide an implementation of `converse` that returns a digraph with all
-/// arcs reversed.
+/// Provide an implementation of [`converse`](Converse::converse) that returns
+/// a digraph with all arcs reversed.
 ///
 /// ```
 /// use {